+++ cloud/ubuntu.sh
@@ -1,7 +1,8 @@
 #! /bin/bash -e
 # Provision an AWS cloud instance with tutor.
-# Run with: curl -sSL https://raw.githubusercontent.com/regisb/tutor/cloud/cloud/ubuntu.sh | sudo bash -e
+# Run with: curl -sSL https://raw.githubusercontent.com/regisb/tutor/master/cloud/ubuntu.sh | sudo bash -e
 
+export TUTOR_VERSION="v3.0.4"
 export USER=ubuntu
 export DEBIAN_FRONTEND=noninteractive 
 
@@ -32,7 +33,7 @@ curl -L "https://github.com/docker/compose/releases/download/1.23.2/docker-compo
 chmod +x /usr/local/bin/docker-compose
 
 echo "=============== Installing tutor"
-curl -L "https://github.com/regisb/tutor/releases/download/v3.0.4/tutor-linux" -o /usr/local/bin/tutor
+curl -L "https://github.com/regisb/tutor/releases/download/$TUTOR_VERSION/tutor-linux" -o /usr/local/bin/tutor
 chmod +x /usr/local/bin/tutor
 
 echo "=============== Configuring supervisor"